Move Announcement struct into separate file
